Make source code checkers accessible from Grunt
To make the development cycle faster we want developers to be able to run
all the required checks locally. Because of the way Composer is designed,
we cannot make the checkers a dev dependencies and have to install them
separately.

Previously we have installed the dependencies manually but keeping them
consistent between CI and local environment would be difficult. For that
reason we created a separate `composer.json` in the `utils` directory.
The incompatible checkers have to be removed from the manifest before
running `composer install` to solve the incompatibilities.

This commit introduces new grunt tasks:

* `lint` will check PHP syntax using built-in PHP tool
* `cs` will check the coding style using PHP-CS-Fixer
* `check` will run both the checkers (more in the future)