Drop alt attribute of source favicons
The favicons in the source list contained the title of the feed in
their alt attribute. This was redundant since the title is already
shown next to the item, resulting in the title being read multiple
times by accessibility technologies. It was also displayed over
the title when the favicon file was missing.

Closes: #975